Make receive/send completion callbacks nothrow
Rationale: Just because an exception occurs in the callback does
not necessarily mean the socket must dies. However, the event loop
cannot make any sensible, generic decision on what to do if an
exception occured, as it does not know about the semantics of
the callbacks body. As such, it seems more reasobale to forbid
exceptions from leaving these callbacks to enforce the user
logic to decide what to do.